0.3.1.1       09/03/2018 Added v4 of YAML package and latest Atoum for compatability with Laravel 5.6 and PHP 7.1
Fixed an issue where trying to coerce a non numeric value into a numeric via math operators throws a warning in php 7.1+
Fix for date validation to use start and end of day for the boundary tests.

+++ src/Helper/Structure.php
@@ -90,7 +90,7 @@ trait Structure {
 
 
   static function addToArray($arr, $val, $key = NULL, $path = NULL) {
-    if (empty($key)) $key = key(array_slice($arr, -1, 1, TRUE)) + 1;
+    if (empty($key)) $key = (int) key(array_slice($arr, -1, 1, TRUE)) + 1;
 
     if (! empty($path)) {
